requirements.txt
discord.py==2.3.2
v2enlib @ git+https://github.com/TakahashiNguyen/MyLib